requirements.txt
Django==1.6.3
Pillow==2.4.0
PyJWT==0.4.1
South==1.0.2
argparse==1.2.1
bleach==1.4
django-appconf==0.6
django-bootstrap-pagination==1.5.0
django-cked==0.1.0
django-debug-toolbar==1.2.2
django-imagekit==3.2.1
django-registration-redux==1.1
future==0.9.0
html5lib==0.999
httplib2==0.9
mock==1.0.1
oauth2==1.5.211
oauthlib==0.6.1
pilkit==1.1.6
python-openid==2.2.5
pytz==2014.2
requests==2.2.1
requests-oauthlib==0.4.0
six==1.6.1
sqlparse==0.1.14
tzlocal==1.0
wsgiref==0.1.2
